chore: reformatting - minor refactoring in function calls
